Feat: Streamlined Rodent species cleaning
The function has an R helper function that can be called from the CLI to extract any of the databases in ArHa, it is exported in UTF-8 encoding and .csv formatting, making it usable in Python.

RodentTaxa.py is the main function, it calls RDS2CSV.R and fixes the names of the rodent species according to GBIF, and then outputs a cleaned database with corrected host species dessignation.

Finally, there's a RodentTaxa_Example.py that provides a usage example for RodentTaxa.py